Code/cloud function/requirements.txt
cloudevents
google-cloud-storage